+++ src/Data/IntMap/Monoidal.hs
@@ -155,6 +155,7 @@ import Data.Semialign (Unalign)
 import Data.Zip (Zip)
 #endif
 #endif
+import qualified Witherable
 
 -- | An 'IntMap' with monoidal accumulation
 newtype MonoidalIntMap a = MonoidalIntMap { getMonoidalIntMap :: M.IntMap a }
@@ -171,6 +172,7 @@ newtype MonoidalIntMap a = MonoidalIntMap { getMonoidalIntMap :: M.IntMap a }
              , Zip
 #endif
 #endif
+             , Witherable.Filterable
              )
 
 deriving instance Eq1 MonoidalIntMap
@@ -242,6 +244,8 @@ instance Semigroup a => IsList.IsList (MonoidalIntMap a) where
     toList = M.toList . unpack
     {-# INLINE toList #-}
 #endif
+
+instance Witherable.Witherable MonoidalIntMap
 
 -- | /O(1)/. A map with a single element.
 singleton :: Int -> a -> MonoidalIntMap a